+++ tiled/auth_database/core.py
@@ -2,15 +2,11 @@ import hashlib
 import uuid as uuid_module
 from datetime import datetime
 
-from alembic import command
-from alembic.config import Config
-from alembic.runtime import migration
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
 from sqlalchemy.orm import selectinload
 from sqlalchemy.sql import func
 
-from .alembic_utils import temp_alembic_ini
 from .base import Base
 from .orm import APIKey, Identity, PendingSession, Principal, Role, Session
 
@@ -66,83 +62,6 @@ async def initialize_database(engine):
         # Initialize Roles table.
         async with AsyncSession(engine) as db:
             await create_default_roles(db)
-
-
-def stamp_head(engine_url):
-    """
-    Upgrade schema to the specified revision.
-    """
-    with temp_alembic_ini(engine_url) as alembic_ini:
-        alembic_cfg = Config(alembic_ini)
-        command.stamp(alembic_cfg, "head")
-
-
-def upgrade(engine_url, revision):
-    """
-    Upgrade schema to the specified revision.
-    """
-    with temp_alembic_ini(engine_url) as alembic_ini:
-        alembic_cfg = Config(alembic_ini)
-        command.upgrade(alembic_cfg, revision)
-
-
-def downgrade(engine_url, revision):
-    """
-    Downgrade schema to the specified revision.
-    """
-    with temp_alembic_ini(engine_url) as alembic_ini:
-        alembic_cfg = Config(alembic_ini)
-        command.downgrade(alembic_cfg, revision)
-
-
-class UnrecognizedDatabase(Exception):
-    pass
-
-
-class UninitializedDatabase(Exception):
-    pass
-
-
-class DatabaseUpgradeNeeded(Exception):
-    pass
-
-
-async def get_current_revision(engine):
-    redacted_url = engine.url._replace(password="[redacted]")
-    async with engine.connect() as conn:
-        context = await conn.run_sync(migration.MigrationContext.configure)
-        heads = await conn.run_sync(lambda conn: context.get_current_heads())
-    if heads == ():
-        return None
-    elif len(heads) != 1:
-        raise UnrecognizedDatabase(
-            f"This database {redacted_url} is stamped with an alembic revisions {heads}. "
-            "It looks like Tiled has been configured to connect to a database "
-            "already populated by some other application (not Tiled) or else "
-            "its database is in a corrupted state."
-        )
-    (revision,) = heads
-    if revision not in ALL_REVISIONS:
-        raise UnrecognizedDatabase(
-            f"The datbase {redacted_url} has an unrecognized revision {revision}. "
-            "It may have been created by a newer version of Tiled."
-        )
-    return revision
-
-
-async def check_database(engine):
-    revision = await get_current_revision(engine)
-    redacted_url = engine.url._replace(password="[redacted]")
-    if revision is None:
-        raise UninitializedDatabase(
-            f"The database {redacted_url} has no revision stamp. It may be empty. "
-            "It can be initialized with `initialize_database(engine)`."
-        )
-    elif revision != REQUIRED_REVISION:
-        raise DatabaseUpgradeNeeded(
-            f"The database {redacted_url} has revision {revision} and "
-            f"needs to be upgraded to revision {REQUIRED_REVISION}."
-        )
 
 
 async def purge_expired(db, cls):
